Separate ZLIB API
This commit separates the function definitions implementing the zlib
interfaces.

The purpose of doing this is to make it easier to build a second version
of the library without the interface symbols.
