+++ lib/nx_adler32.c
@@ -172,40 +172,3 @@ unsigned long nx_adler32_combine(adler1, adler2, len2)
     if (sum2 >= BASE) sum2 -= BASE;
     return sum1 | (sum2 << 16);
 }
-
-/* ========================================================================= */
-
-#ifdef ZLIB_API
-unsigned long adler32_combine(adler1, adler2, len2)
-    unsigned long adler1;
-    unsigned long adler2;
-    z_off_t len2;
-{
-    return nx_adler32_combine(adler1, adler2, len2);
-}
-
-unsigned long adler32_combine64(adler1, adler2, len2)
-    unsigned long adler1;
-    unsigned long adler2;
-    z_off64_t len2;
-{
-    return nx_adler32_combine(adler1, adler2, len2);
-}
-
-unsigned long adler32(adler, buf, len)
-    unsigned long adler;
-    const char *buf;
-    unsigned int len;
-{
-    return nx_adler32_z(adler, buf, len);
-}
-
-unsigned long adler32_z(adler, buf, len)
-    unsigned long adler;
-    const char *buf;
-    z_size_t len;
-{
-    return nx_adler32_z(adler, buf, len);
-}
-#endif
-
