+++ lib/nx_crc.c
@@ -473,30 +473,3 @@ uLong nx_crc32_combine64(crc1, crc2, len2)
 {
     return crc32_combine_(crc1, crc2, len2);
 }
-
-#ifdef ZLIB_API
-unsigned long crc32(crc, buf, len)
-    unsigned long crc;
-    const unsigned char FAR *buf;
-    uint64_t len;
-{
-    return crc32_ppc(crc, (unsigned char *)buf, len);
-}
-
-uLong crc32_combine(crc1, crc2, len2)
-    uLong crc1;
-    uLong crc2;
-    uint64_t len2;
-{
-    return nx_crc32_combine(crc1, crc2, len2);
-}
-
-uLong crc32_combine64(crc1, crc2, len2)
-    uLong crc1;
-    uLong crc2;
-    uint64_t len2;
-{
-    return nx_crc32_combine64(crc1, crc2, len2);
-}
-#endif
-
