+++ lib/nx_deflate.c
@@ -2395,54 +2395,3 @@ mem_error:
 
 	return Z_MEM_ERROR;
 }
-
-
-#ifdef ZLIB_API
-int deflateInit_(z_streamp strm, int level, const char* version, int stream_size)
-{
-	return nx_deflateInit_(strm, level, version, stream_size);
-}
-
-int deflateInit2_(z_streamp strm, int level, int method, int windowBits,
-		int memLevel, int strategy, const char *version,
-		int stream_size)
-{
-	return nx_deflateInit2_(strm, level, method, windowBits, memLevel, strategy, version, stream_size);
-}
-
-int deflateReset(z_streamp strm)
-{
-	return nx_deflateReset(strm);
-}
-
-int deflateEnd(z_streamp strm)
-{
-	return nx_deflateEnd(strm);
-}
-
-int deflate(z_streamp strm, int flush)
-{
-	return nx_deflate(strm, flush);
-}
-
-unsigned long deflateBound(z_streamp strm, unsigned long sourceLen)
-{
-	return nx_deflateBound(strm, sourceLen);
-}
-
-int deflateSetHeader(z_streamp strm, gz_headerp head)
-{
-	return nx_deflateSetHeader(strm, head);
-}
-
-int deflateSetDictionary(z_streamp strm, const Bytef *dictionary, uInt  dictLength)
-{
-	return nx_deflateSetDictionary(strm, dictionary, dictLength);
-}
-
-int deflateCopy(z_streamp dest, z_streamp source)
-{
-	return nx_deflateCopy(dest, source);
-}
-
-#endif
