+++ lib/nx_inflate.c
@@ -1747,42 +1747,3 @@ int nx_inflateGetHeader(z_streamp strm, gz_headerp head)
 
 	return Z_OK;
 }
-
-
-#ifdef ZLIB_API
-int inflateInit_(z_streamp strm, const char *version, int stream_size)
-{
-	return nx_inflateInit_(strm, version, stream_size);
-}
-int inflateInit2_(z_streamp strm, int windowBits, const char *version, int stream_size)
-{
-	return nx_inflateInit2_(strm, windowBits, version, stream_size);
-}
-int inflateReset(z_streamp strm)
-{
-	return nx_inflateReset(strm);
-}
-int inflateEnd(z_streamp strm)
-{
-	return nx_inflateEnd(strm);
-}
-int inflate(z_streamp strm, int flush)
-{
-	return nx_inflate(strm, flush);
-}
-int inflateSetDictionary(z_streamp strm, const Bytef *dictionary, uInt dictLength)
-{
-	return nx_inflateSetDictionary(strm, dictionary, dictLength);
-}
-
-int inflateCopy(z_streamp dest, z_streamp source)
-{
-	return nx_inflateCopy(dest, source);
-}
-
-int inflateGetHeader(z_streamp strm, gz_headerp head)
-{
-	return nx_inflateGetHeader(strm, head);
-}
-
-#endif
