+++ lib/nx_uncompr.c
@@ -72,17 +72,3 @@ int nx_uncompress(Bytef *dest, uLongf *destLen, const Bytef *source, uLong sourc
 {
     return nx_uncompress2(dest, destLen, source, &sourceLen);
 }
-
-#ifdef ZLIB_API
-
-int uncompress(Bytef *dest, uLongf *destLen, const Bytef *source, uLong sourceLen)
-{
-	return nx_uncompress(dest, destLen, source, sourceLen);
-}
-int uncompress2(Bytef *dest, uLongf *destLen, const Bytef *source, uLong *sourceLen)
-{
-	return nx_uncompress2(dest, destLen, source, sourceLen);
-}
-
-#endif
-
